Add InputFilesReader class
This class can serve as a wrapper for the specific file readers, as well as a main method we can draw on to create the population file. As a note, the readActivitySimFiles method could be called directly from RunMatsim if we wanted to skip creating the population xml. But as of this commit, the main method creates a population xml from the activitysim input files.